requirements.txt
alembic==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
bcrypt==3.2.2
blinker==1.5
CacheControl==0.12.11
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
crypto==1.4.1
cryptography==37.0.2
decorator==5.1.1
Deprecated==1.2.13
dnspython==1.15.0
email-validator==1.2.1
firebase-admin==5.2.0
Flask==2.0.3
flask-blueprint==1.3.0
flask-csp==0.10
Flask-JWT-Extended==4.4.2
Flask-Limiter==2.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
gcloud==0.18.3
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.64.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.3
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-kms==2.12.0
google-cloud-logging==3.2.1
google-cloud-secret-manager==2.12.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3
limits==2.6.3
Mako==1.2.1
MarkupSafe==2.1.1
mkdocs-material-extensions==1.0.3
msgpack==1.0.4
Naked==0.1.31
oauth2client==4.1.3
oauthlib==3.2.1
packaging==21.3
proto-plus==1.22.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.15.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-json-logger==2.0.4
python-jwt==3.3.2
pytz==2022.2.1
PyYAML==6.0
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
rsa==4.9
schedule==1.1.0
sentry-sdk==1.9.4
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.40
stripe==3.5.0
tabulate==0.8.10
twilio==7.12.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.1.2
wget==3.2
wrapt==1.14.1
WTForms==2.3.3
yarg==0.1.9
zipp==3.8.0
